Change pusher to not touch the original app
We changed the way the blue/green occurs. The new process is to push a
new version of the application with a temporary name. If the deploy is
successful it will delete the original application and rename it to not
have the temporary name.

If the deploy is unsuccessful, it will only delete the temporary
applicaiton that was pushed.

- Rename Rollback to UndoPush

[finish #138503947]

Signed-off-by: Travis Yoder <trayoda@gmail.com>